backend/requirements.txt
asgiref==3.6.0
sqlparse==0.5.3
Django==4.2            # Version stable de Django
djangorestframework==3.14.0
psycopg2-binary==2.9.7  # Driver pour PostgreSQL
gunicorn==21.1.0       # Serveur WSGI pour production
python-dotenv==1.0.0   # Gestion des variables d'environnement (optionnel)
dj-database-url==2.3.0  # Pour gérer les URLs de base de données (ajoutez une version si possible)
djangorestframework-simplejwt==4.7.2 # Gestion JWT
django-cors-headers==3.12.0 # Pour gerer les problemes de CORS
dj-rest-auth==2.1.9
django-allauth==0.54.0
django-extensions==3.2.3 # Afficher les urls du projets
drf-yasg==1.21.4 # Pour swagger
Pillow>=9.4.0,<11.0.0 # pour l'avatar